+++ tests/features/test_illumina.py
@@ -9,12 +9,57 @@ Created on Mon May  3 17:09:17 2021
 import pathlib
 import unittest
 import types
+import datetime
 
 from src.features.illumina import (
-    read_snpMap, read_Manifest, read_snpList, read_illuminaRow)
+    read_snpMap, read_Manifest, read_snpList, read_illuminaRow, skip_lines,
+    skip_until_section, search_manifactured_date)
 
 FEATURE_DATA_DIR = pathlib.Path(__file__).parents[1] / "features/data"
 SCRIPTS_DATA_DIR = pathlib.Path(__file__).parents[1] / "data/data"
+
+
+class SkipTest(unittest.TestCase):
+    data_path = SCRIPTS_DATA_DIR / "test_manifest.csv"
+
+    def test_skip_lines(self):
+        reference = [
+            "Illumina, Inc.,,,,,,,,,,,,,,,,,,,",
+            "[Heading],,,,,,,,,,,,,,,,,,,,",
+            "Descriptor File Name,OvineSNP50_B.bpm,,,,,,,,,,,,,,,,,,,",
+            "Assay Format,Infinium HD Ultra,,,,,,,,,,,,,,,,,,,",
+            "Date Manufactured,1/7/2009,,,,,,,,,,,,,,,,,,,",
+            "Loci Count ,54241,,,,,,,,,,,,,,,,,,,",
+            "[Assay],,,,,,,,,,,,,,,,,,,,",
+        ]
+        with open(self.data_path) as handle:
+            _, skipped = skip_lines(handle, skip=7)
+
+        self.assertListEqual(skipped, reference)
+
+    def test_skip_until_section(self):
+        reference = [
+            "Illumina, Inc.,,,,,,,,,,,,,,,,,,,",
+            "[Heading],,,,,,,,,,,,,,,,,,,,",
+            "Descriptor File Name,OvineSNP50_B.bpm,,,,,,,,,,,,,,,,,,,",
+            "Assay Format,Infinium HD Ultra,,,,,,,,,,,,,,,,,,,",
+            "Date Manufactured,1/7/2009,,,,,,,,,,,,,,,,,,,",
+            "Loci Count ,54241,,,,,,,,,,,,,,,,,,,",
+            "[Assay],,,,,,,,,,,,,,,,,,,,",
+        ]
+        with open(self.data_path) as handle:
+            _, skipped = skip_until_section(handle, "[Assay]")
+
+        self.assertListEqual(skipped, reference)
+
+    def test_search_manifactured_date(self):
+        with open(self.data_path) as handle:
+            _, skipped = skip_until_section(handle, "[Assay]")
+
+        test = search_manifactured_date(skipped)
+        reference = datetime.datetime(2009, 1, 7)
+
+        self.assertEqual(reference, test)
 
 
 class IlluminaMixin():